+++ src/backend/mailbox_manager/tasks.py
@@ -2,6 +2,8 @@
 
 import time
 
+from django.conf import settings
+
 import requests
 from celery import Celery
 from celery.schedules import crontab
@@ -54,12 +56,13 @@ def fetch_domains_status_task(status: str):
     for domain in MailDomain.objects.filter(status=status):
         old_status = domain.status
         # wait 10 seconds between each domain treatment to avoid overloading dimail
-        time.sleep(10)
+        time.sleep(settings.MAIL_CHECK_DOMAIN_INTERVAL)
         try:
             client.fetch_domain_status(domain)
         except requests.exceptions.HTTPError as err:
             logger.error("Failed to fetch status for domain %s: %s", domain.name, err)
         else:
             if old_status != domain.status:
+                domain.notify_status_change()
                 changed_domains.append(f"{domain.name} ({domain.status})")
     return changed_domains